Made some progress. Yeah. Tired.
Basically we have one side of the erased rows going to the end of the
screen.
I fixed some of the erasing problems when we move verticle only to
discover more when we erase with some aliens missing. I think the
solution will be just to erase the entire block when we move down
vertically rather than erase individual aliens like we havve been doing.

I know how to also implement the left side, but I'm so tired that I'm
coding slowly and taking forever to actually think through problems.
It's freaking 2:30 am. I'm calling quits for the night.